# db: require validity check before calling HasPointAndRange, RangeKeyChanged